Refactors the smt code to allow for some code reuse with the
role abstraction inference and network compression.

+++ projects/batfish-common-protocol/src/main/java/org/batfish/datamodel/StaticRoute.java
@@ -60,8 +60,8 @@ public class StaticRoute extends AbstractRoute {
 
   @Override
   @JsonIgnore
-  public Integer getMetric() {
-    return 0;
+  public Long getMetric() {
+    return 0L;
   }
 
   @Nonnull
